[local-preview] Add Windows and Mac support
This PR adds steps to make the `local-preview` work for Windows and
Mac. As the same Windows and Mac steps work for Linux too, For Consistency
purposes This is the default path now.

Work will be done before changing this to `beta` to automate even
the above Private IP retrieval steps.

Signed-off-by: Tarun Pothulapati <tarun@gitpod.io>